+++ R/cv.nc.hdsvm.R
@@ -13,6 +13,10 @@
 #' @param foldid Optional vector of values between 1 and \code{nfolds}
 #'   identifying the fold of each observation. If provided, it overrides
 #'   \code{nfolds}.
+#' @param ncores Number of processes used to fit the folds in parallel
+#'   (with the \pkg{parallel} package: forked processes on Unix-alikes, a
+#'   socket cluster on Windows). Default is 1, i.e. the folds are fitted
+#'   sequentially.
 #' @param ... Additional arguments passed to \code{\link{nc.hdsvm}}
 #'   (e.g. \code{pen}, \code{aval}, \code{lam2}, \code{lla_step}).
 #'
@@ -62,7 +66,7 @@
 #'                          pen = "scad")
 #' cv.nc.fit
 #' }
-cv.nc.hdsvm <- function(x, y, lambda = NULL, nfolds = 5L, foldid, ...) {
+cv.nc.hdsvm <- function(x, y, lambda = NULL, nfolds = 5L, foldid, ncores = 1L, ...) {
   ####################################################################
   ## data setup
   y <- drop(y)
@@ -78,13 +82,13 @@ cv.nc.hdsvm <- function(x, y, lambda = NULL, nfolds = 5L, foldid, ...) {
     foldid <- sample(rep(seq(nfolds), length = x.row)) else nfolds <- max(foldid)
   if (nfolds < 3)
     stop("nfolds must be at least 3; nfolds = 5 recommended.")
-  outlist <- as.list(seq(nfolds))
   ## fit the model nfold times and save them
-  for (i in seq(nfolds)) {
+  dots <- list(...)
+  outlist <- cv_folds(nfolds, function(i) {
     which <- foldid == i
-    outlist[[i]] <- nc.hdsvm(x = x[!which, , drop = FALSE],
-                             y = y[!which], lambda = lambda, ...)
-  }
+    do.call(nc.hdsvm, c(list(x = x[!which, , drop = FALSE], y = y[!which],
+                             lambda = lambda), dots))
+  }, ncores)
   ## select the lambda according to predmat
   cvstuff <- cvpath.hdsvm(outlist, x, y, lambda, foldid, x.row, ...)
   cvm <- cvstuff$cvm